% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{locations}
\format{A \code{data.frame} with 10100 observations and 14 columns. The columns are defined as follows:
\describe{
 \item{\code{track_id}}{unique ID for each track}
}}
\source{
Million Songs Dataset
}
\usage{
locations
}
\description{
This data set contains a subset of song data
}
\examples{
\dontrun{
data("locations")
summary(locations)
}
}
\author{
JB
}
\keyword{datasets}
