% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singer_locations.R
\docType{data}
\name{singer_locations}
\alias{singer_locations}
\title{singer locations}
\format{A \code{data.frame} with 873,684 observations and 6 columns. The columns are defined as follows:
\describe{
 \item{\code{column}}{column definition}
}}
\source{
Million Songs Dataset
}
\usage{
singer_locations
}
\description{
This data set contains song data and location data
}
\examples{
\dontrun{
data("singer_locations")
summary(singer_locations)
}
}
\author{
JB
}
\keyword{datasets}
