% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singer_locations.R
\docType{data}
\name{singer_locations}
\alias{singer_locations}
\title{singer locations}
\format{A \code{data.frame} with 10100 observations and 14 columns. The columns are defined as follows:
\describe{
 \item{\code{track_id}}{unique ID for each track}
  \item{\code{duration}}{duration of the track in seconds}
  \item{\code{release}}{album name from which the track was taken, some songs / tracks can come from many albums, only one is given}
}}
\source{
Million Songs Dataset
}
\usage{
singer_locations
}
\description{
This data set contains song data and location data
}
\examples{
\dontrun{
data("singer_locations")
summary(singer_locations)
}
}
\author{
JB
}
\keyword{datasets}
