% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{songs}
\alias{songs}
\title{songs}
\format{A \code{data.frame} with 22 observations and 3 columns. The columns are defined as follows:
\describe{
 \item{\code{artist_name}}{artist name}
 \item{\code{title}}{song title}
 \item{\code{year}}{year when the song was released, according to musicbrainz.org}
}}
\source{
\url{http://labrosa.ee.columbia.edu/millionsong/}
}
\usage{
songs
}
\description{
This data set contains a small subset of song data from the Million Song Dataset including the artist name, song title and year of release.
}
\examples{
\dontrun{
data("songs")
summary(songs)
}
}
\author{
Joey Bernhardt
}
\keyword{datasets}
